requirements.txt
numpy==1.26.4
pillow==10.2.0
pyaml==23.12.0
PyYAML==6.0.1

[sdxl]

[kandinsky]